change key type from u8 to u32

+++ kernel/include/sys/tty.h
@@ -1,4 +1,4 @@
 #ifndef _DIYOS_TTY_H
 #define _DIYOS_TTY_H
-extern void show(u8 key);
+extern void show(u32 key);
 #endif
